Enhanced scan (#10)
scan returns 10 results, and accepts "it" as a command to ITerate
through another 10 results. Iteration is allowed until either we reach
the end of results OR a command other than "it" is used (this destroys
the cursor).

scan also supports an optional parameter which is the prefix to limit
our scan to.