Reorganize and treat commands as first class objects
This is an attempt at refactoring code and having commands more
self-contained. I only did "get" so that we could discuss it and decide
if we want to move forward.

The idea is to have commands more self-contained, but able to use common
functionality (such as parsing the input, and whatever context
provides). Having a command-per-file is a little easier to follow, and
instead of having input parsed to an arbitrary number of values (which
is hard in a static language), the parsing and the execution happen
together.

This would also make it easier to more completely unit tests commands,
but I find that the current withinShell helper is probably still the
right approach (there's very little overhead and running through the
entire process)